Catalyst.Step()
Gave Catalyst a Step() method and made ProgressiveCatalyst use it to increase progress instead of increasing progress in Catalyze().

Also added Step() to Cell and Slot to enable stepping all the catalysts by simply calling Cell.Step(). (This has added benefit of allowing cells and slots to respond to time in other ways in the future).

Decoupling the production of actions from responding to the passage of time is a natural change, and makes future changes to Catalyst functionality less difficult. (Syncronization between catalysts during Catalyze() is impossible without separate time function, for one)