Separatase
Added new Catalyst Separatase. This catalyst uses ATP to break an organism into two pieces, as well as sending "seed" ATP to the new organism ("new" means the one that doesn't contain the Separatase)

In order to acheive this effect, Organism.SeparateCell() was renamed to just "Separate" and its functionality was completely rewritten.