Membrane.GetTransportRate() changes
Membrane.GetTransportRate() now takes the contents slots into consideration when determining the concentration of the Cytosol. This models the real process, and the original intent for how pumping/ interacting with WaterLocale was supposed to work; the rate of transfer across the membrane is proportional to the concentration inside the cell and outside. Because compounds are no longer stored in the cytosol, without this change, the rate would always be the max rate (because concentration inside would be zero).

I also changed the formula to try and smooth out the rate changes that occur (0 concentration -> 0.001 concentration is still infinite concentration gradient -> finite). I now add a very small value to each to concentration when calculating the gradient (you can think of this as approximating some sort of "friction"), and I adapted the formula from here http://www.nimbios.org/~gross/bioed/webmodules/diffusion.htm for active transport. This new formula has a built in maximum, and even better, approaches / falls way from this maximum smoothly.

Still, huge swings in the gradient mess with this smoothness, so I decided to make smoles (the game's smaller version a "mole") 10x smaller, so in effect, 1 unit makes 10x less difference in concentration.