Catalyst.IsSame()
Catalyst.IsSame() is a function that compares two catalysts, and returns true if they are the same molecule. As in, if in the real world, I have two identical dna strands, this method would consider them "same". They exist in different places, so they aren't "equal", and one might have a transciption protein attached to it, so they aren't "stackable".

This is necessary for identifying Ribozymes and Enzymes by their underlying Catalyst. e.g. The list of Ribozymes shouldn't be cluttered with every single rotation of a given catalyst, or every possible cofactor if could have.

In order of similarity, these are the current comparisons:
== (Same reference)
.Equals() (Same everything else; used by Undo)
.IsStackable() (Similar enough to be stacked together)
.IsSame() (Just the same molecule, structurally)