ProgressiveCatalyst.GetNormalizedClaimYield() bugfix
This method used to wholly remove non-InstantCatalyst's claims if their resource had to be shared. This had the unintended effect of making their claim yield one, when it should be zero. Now, they are just added to the list of claimants who have been processed, and their claim yield is set to zero.

Also added some comments to help explain the algorithm better.