InstantCatalyst bugfixes
InstantCatalyst now checks whether an Action is in the current Stage, and whether it IsLegal in Communicate().

It also now checks for IsLegal in Catalyze(), and checks if the NormalizedClaimYield is zero, in which case it returns null.

These changes prevent mis-staged, illegal, and pointless actions from progressing to further steps.