Interpretase.LoadProgram change
Interpretase.LoadProgram now adds an "LLL" codon to the beginning of the program when it finishes loading it.

This is necessary to allow user to to copy their entire DNA strand, because if they put their own locuses around the program, the CopyCommand is exclusive, so the locuses themselves don't get copied.