NRG changes
Chargeable NRG was saving with name "Unnamed" because when Chargeable attempts to save the NRG SimpleMolecule, the system can't find a matching named SimpleMolecule because that name ("NRG") now refers to the ChargeableMolecule.

To fix this, now "NRG" always refers to the SimpleMolecule, and new names "Charged NRG" and "Discharged NRG" refer to the charged and discharged versions of the ChargeableMolecule.

This also has benefit of allowing reaction definitions to specifiy charged or uncharged NRGs as reactants or products.