"Enzyme" -> "Protein"
Changing terminology. "Enzyme" was poorly named; a catalytic molecule made up of amino acids _is_ an enzyme, but so is a ribozyme (which my code makes a distinction between). The exact word is "protein", so I've gone ahead and changed this.

In line with this, I considered changing "Catalyst" to the more specific word "Enzyme", thereby making Ribozymes and Proteins both Enzymes. However, I don't want to rule out the possibility of having a non-enzymatic catalyst, so I'm leaving that name as is for now.