Action scaling changes
Action now has a BaseScale and a StackScale to compute final Scale. This is necessary to allow the new feature of stacks requiring more and more effort to increase their size. StackScale is a functional value, calculated based on the stacking requirements scaled by BaseScale.

Due to this change, scale can't be interacted with directly anymore because of the *= operator which would fold the adjusted Scale back in rather than just change BaseScale. So now, Scale has to be adjusted using ScaleByFactor().