Data based game balance
This change adds two new files: Balance.cs and the data file that drives it, balance.json.

These files are intended to connect the arbitrary numbers associated with the implementation of all game elements to a single human readable file. For example, does cell construction require 4 units of structate, or 5? Does it consume 1 smole or NRG, or 0.75? etc.

This enables fast iteration of low-code-impact gameplay changes, as well as limited modding potential. It also means that these sort of changes are centralized to one file, instead of being scattered across the entire codebase.

This commit does not actually make the relevant classes use Balance.cs. I'm breaking that up into separate commits.