Interpretase CopyCommand balance integration
This process is one of the main reasons the balance file was concieved - With the introduction of CopyCommand, and later, its redesign to get rid of holdover design from CutCommand, its function needed to be related to the function of Transcriptase numerically. Instead of just having them share notes, I decided to futureproof this sort of thing by creating a central file describing the numerical function of all game elements.

As such, this commit is tangled with the changes I was making to CopyCommand to improve its design.