Measures.GetVisualQuantity()
This function allows effectively changing the units displayed on screen for various compounds.

More specifically, it scales the quantity displayed for NRG by 10x, and for polymers 100x. This is necessary to have realistic resource costs without having wacky quantities displayed.

For example, without this change, 1 unit of DNA might require 100 units of nucleotides, or I'd just have to display 0.01 units of DNA. One is a balance problem, the other, a visualization problem.