MoveAction.IsLegal changes
MoveAction now checks whether the original quantity demanded is available and .IsLegal is false if not.

This makes "take" commands fail if they can't grab the specified amount of resources, and thus interpretases will wait until they are available.

This does add failure points to user's programs, but the alternative is they have to make much more complex "wait" commands for what is likely a very common use case.