Catalyst.ClearState()
This new method is for getting a fresh copy of a Catalyst, so that, for example, you can copy interpretase without copying the dna program that particular instance contains.

This feature is also a bugfix; that particular example was happening in Interpretase.CopyCommand.

Organism.GetDeck() also uses this feature to make fresh copies for saving (really just adds human readability / reduces file size).