New Utility functions
Added CreateLookup(), CreateInverseLookup() and GetEnumValues() to Utility.

The first function allow you to bake a lookup function given a System.Func and a IEnumerable domain. The second allows you to do the same, but inverted; a->b becomes b->a.

The last function is just a more helpful version of System.Enum.GetValues(), which isn't IEnumerable for some reason.

The use case in mind here is to allow me to have a single function mapping enum values to strings, and then I automatically have a function that maps strings back to enum values.