Program.cs
I've completely replaced the earlier system of having DNA contain a description of sectors, and having the visual layer try to keep synced with the DNA.

Instead, DNA is now just a information medium, the program itself exists within the new class "Program". Program does away with codons and nucleotides, and focuses plainly on representing the tokens and organization structures present in a program.

Each program contains a single Sector, which itself contains a list of "Code"s. Codes come in two types: Sectors and Tokens. Tokens describe the function of the program, sectors are purely organization. There are 4 types of Tokens; CommandToken, LocusToken, ValueToken and FunctionToken. Tokens all have a unique codon representation, thus, a program can be translated into a DNA sequence.

Programs are stored in a list within Organism, and can be looked up by DNA sequence. In this way, when a user wishes to edit a DNA strand, the matching program can be found and displayed on screen.

Because this representation is hierarchical, which matches the visualization, and because it is designed from the ground up to be edited, now the visual layer can keep synced with the logical layer (the Program) with a minimum opportunity for bugs.

When a program is displayed visually, each visual element has a direct reference to the Program element it represents, and thus any change can be very simply reflected in the Program. For example, if you change a LocusNode's value to "LLV", you just have to change the LocusTokens value to the corresponding number, and you're done.

I've also taken the opportunity to improve the software design of the visual elements. I tried to make them more independent, both of the the logical layer, as well as eachother. For example, before, you had to know what a DNAPanelNode's CodonIndex and CodonLength were in order to apply changes to the DNA. Now, because a LocusNode simply has a reference to an instance of LocusToken, edits can either look like simply editing the value of a LocusToken, or, if your need to remove the Token, you can say Sector.RemoveCode(locus_token), and you're done! What this means is that all this tight coupling required to keep track of aggregate information is gone. Nodes and Codes can all just exist without any concept of relative values like position.

Finally, when changes are made to the program via the visual layer, DNAPanel detects them and applies the changes the to the DNA by simply clearing the DNA and rebuilding it with the new sequence, eliminating the error prone process of carefully excising and inserting sequences.

All of the code for interpreting the information in a program still operates on DNA. In future commits, I'd like to move this functionality into Program, so that f.e. You can query the number of operands to a command without having to convert the program to DNA and then query Interpretase.