CatalystProgressIcon
Added files to support an icon that visualizes the process of completing a long action.

Before, the user couldn't tell anything was happening when a multi-turn action was being worked on. Now, a little wheel/pie appears and grows from a little red slice to a full green circle. Based on pikmin's progress wheel.

In order to suppor this, MoleculeComponent was updated to contain this wheel and move it around to keep it away from the quantity labels in screen space.