ActionAnimation bugfix
ActionAnimation was not destroying itself if played in reverse, resulting in rogue animation scripts running around.

While fixing it I triggered another hidden bug, which was the fact that GetMoment() was effecting the state of smooth_moment. Fixing this, I discovered how truly truly broken Unity's SmoothDamp function is (or I suppose its possible I'm just really dumb).

I thought it simply failed to get 100% of the way to the target value in the specified time. However, I was actually updating it twice a frame; turns out normally it barely makes it past halfway... I even tested this a completely separate script.

For now, I'm just giving the function half the real time and then scaling smooth_moment by ~1.1 to get the rest of the way (Even with 4x more than the target time, it won't make it, enough to make a ugly snap-into-place when the animation completes)