UIOverlay redesign
I redesigned this class because I intended to draw the new CatalystProgressIcons through it, however, animating the object was much easier when it was part of the scene, so I removed the new code related to that from UIOverlay.

The changes to support this feature were good however, so I'm leaving them in.