Separation animation improvement
The separation animation now includes _all_ separated cells, and moves away from the new center of mass (of parent organism) as opposed to always moving toward North.

Had to reorganize Organism's separation code to expose parts of it to the animation.

Also added several overrides to MoveAnimation.SetParameters() to allow the use of fixed positions for source and target.