MoleculeComponent changes
Updated MoleculeComponent and its prefab to allow for more advanced animation. In particular, animating the molecule icon.

Changes include giving MoleculeComponent a background image. This is helpful for catalysts because they all share a common background ring.

I moved the main sprite into the prefab's hierarchy as a child to allow drawing the main sprite over another sprite.

I added two sprite layers above the main sprite to allow for animated components on top of the main sprite.

I also updated many sprite layer order values to make sure everything renders in the correct sequence.

During rebase, this commit got merged with the next commit, which added a
unique image for Constructase, and animated it using the foundations
set down in this commit. These commits are far to tricky for me to
bother resolving with my limited git-foo, so I'm leaving them merged.